Fix warnings between floats and uint64

+++ util.c
@@ -1018,7 +1018,7 @@ parse_size_with_suffix(const char *str, uint64_t *size)
 		// Default suffix: G.
 		x *= 1000 * 1000 * 1000;
 	}
-	*size = x;
+	*size = (uint64_t)x;
 	return true;
 }
